[Backend] fix: Resolve circular dependencies in test files Redefining the _test files to eliminate circular dependencies and refactored to accommodate tests within their respective packages.
